only kill users that have a valid login shell

+++ scripts/linux/kill_all_users.sh
@@ -12,8 +12,15 @@ w
 
 for user in $userlist
 do
-	echo "Attempting to kill user $user..."
-	pkill -KILL -u $user
+	usershell=$(cat /etc/passwd | grep $user | cut -d : -f 7)
+	for shell in $(cat /etc/shells)
+	do
+		if [ "$usershell" = "$shell" ]
+		then 
+			echo "Attempting to kill user $user..."
+			pkill -KILL -u $user
+		fi
+	done
 done
 
 echo "Waiting for processes to end..."
